Migrate getStatus
For some reason I can’t load the videos page (e.g.
https://video.local.dev-gutools.co.uk/videos/ebdea793-16cf-45eb-bf84-acf52ccd0a0d),
and get an error “this.props.workflowActions.getStatus is not a
function” coming from (I think) the use in pages/Video/index.jsx.

I can’t work out what I’ve missed by comparing with the youtube or path
reducers.
