+++ public/video-ui/src/components/Workflow/Workflow.jsx
@@ -63,8 +63,7 @@ class Workflow extends React.Component {
 
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as getStatus from '../../actions/WorkflowActions/getStatus';
-import {getSections} from '../../slices/workflow';
+import { getStatus, getSections } from '../../slices/workflow';
 import * as getStatuses from '../../actions/WorkflowActions/getStatuses';
 import * as getPriorities from '../../actions/WorkflowActions/getPriorities';
 import * as localUpdateWorkflowData from '../../actions/WorkflowActions/localUpdateWorkflowData';
