+++ public/video-ui/src/reducers/workflowReducer.js
@@ -1,9 +1,5 @@
 export default function workflow(state = { sections: [], statuses: [], status: {}, priorities: [] }, action) {
   switch (action.type) {
-    case 'WORKFLOW_SECTIONS_GET_RECEIVE':
-      return Object.assign({}, state, {
-        sections: action.sections || []
-      });
     case 'WORKFLOW_STATUSES_GET_RECEIVE':
       return Object.assign({}, state, {
         statuses: action.statuses || []
@@ -11,10 +7,6 @@ export default function workflow(state = { sections: [], statuses: [], status: {
     case 'WORKFLOW_PRIORITIES_GET_RECEIVE':
       return Object.assign({}, state, {
         priorities: action.priorities || []
-      });
-    case 'WORKFLOW_STATUS_GET_RECEIVE':
-      return Object.assign({}, state, {
-        status: Object.assign({}, {isTrackedInWorkflow: true}, action.status) || {}
       });
     case 'WORKFLOW_STATUS_NOT_FOUND':
       return Object.assign({}, state, {
